fix issue #3623: Renaming key to one that exists
The tab index is set to the next tab if error occurs.
All tabs are missing if the focus is on the last tab, because
tabs.currentIndex is set to a tab that does not exist.
If you activate a tab other than the last one, when error occurs, the
next tab is activated instead of the current one.